+++ src/chains/supportedChains.evm.ts
@@ -1713,4 +1713,52 @@ export const supportedEVMChains: EVMChain[] = [
       rpcUrls: ['https://rpc.sophon.xyz'],
     },
   },
+
+  {
+    key: ChainKey.PLA,
+    chainType: ChainType.EVM,
+    name: 'Plasma',
+    coin: CoinKey.XPL,
+    id: ChainId.PLA,
+    mainnet: true,
+    logoURI:
+      'https://raw.githubusercontent.com/lifinance/types/main/src/assets/icons/chains/plasma.svg',
+    relayerSupported: false,
+    multicallAddress: multicallAddresses[ChainId.PLA],
+    metamask: {
+      chainId: prefixChainId(ChainId.PLA),
+      blockExplorerUrls: ['https://plasmascan.to/'],
+      chainName: 'Plasma',
+      nativeCurrency: {
+        name: CoinKey.XPL,
+        symbol: CoinKey.XPL,
+        decimals: 18,
+      },
+      rpcUrls: ['https://rpc.plasma.to'],
+    },
+  },
+
+  {
+    key: ChainKey.FLW,
+    chainType: ChainType.EVM,
+    name: 'Flow',
+    coin: CoinKey.FLOW,
+    id: ChainId.FLW,
+    mainnet: true,
+    logoURI:
+      'https://raw.githubusercontent.com/lifinance/types/main/src/assets/icons/chains/flow.svg',
+    relayerSupported: false,
+    multicallAddress: multicallAddresses[ChainId.FLW],
+    metamask: {
+      chainId: prefixChainId(ChainId.FLW),
+      blockExplorerUrls: ['https://evm.flowscan.io/'],
+      chainName: 'Flow EVM Mainnet',
+      nativeCurrency: {
+        name: CoinKey.FLOW,
+        symbol: CoinKey.FLOW,
+        decimals: 18,
+      },
+      rpcUrls: ['https://mainnet.evm.nodes.onflow.org'],
+    },
+  },
 ]